test(gateway): move Linux native admission coverage to a dedicated test file
