+++ src/gateway/server/ws-connection/handshake-auth-helpers.test.ts
@@ -7,7 +7,6 @@ import {
 import type { ConnectParams } from "../../../../packages/gateway-protocol/src/schema.js";
 import type { AuthRateLimiter } from "../../auth-rate-limit.js";
 import {
-  isNativeAppUiClient,
   resolveHandshakeBrowserSecurityContext,
   resolvePairingLocality,
   resolveUnauthorizedHandshakeContext,
@@ -68,13 +67,6 @@ const CLI_CONNECT_PARAMS = {
   client: {
     id: GATEWAY_CLIENT_IDS.CLI,
     mode: GATEWAY_CLIENT_MODES.CLI,
-  },
-} as ConnectParams;
-
-const LINUX_NATIVE_CONNECT_PARAMS = {
-  client: {
-    id: GATEWAY_CLIENT_IDS.LINUX_APP,
-    mode: GATEWAY_CLIENT_MODES.UI,
   },
 } as ConnectParams;
 
@@ -152,51 +144,6 @@ function preserveLocalCliSharedAuthScopes(overrides: Partial<LocalCliSharedAuthS
 }
 
 describe("handshake auth helpers", () => {
-  it("auto-approves local openclaw-linux pairing and keeps remote pairing explicit", () => {
-    expect(isNativeAppUiClient(LINUX_NATIVE_CONNECT_PARAMS.client)).toBe(true);
-    const locality = resolvePairingLocality({
-      connectParams: LINUX_NATIVE_CONNECT_PARAMS,
-      isLocalClient: true,
-      requestHost: "127.0.0.1:18789",
-      remoteAddress: "127.0.0.1",
-      hasProxyHeaders: false,
-      hasBrowserOriginHeader: false,
-      sharedAuthOk: true,
-      authMethod: "token",
-    });
-    expect(locality).toBe("direct_local");
-    expect(
-      shouldAllowSilentLocalPairing({
-        locality,
-        hasBrowserOriginHeader: false,
-        isControlUi: false,
-        isWebchat: false,
-        isNativeAppUi: true,
-        reason: "not-paired",
-      }),
-    ).toBe(true);
-    expect(
-      shouldAllowSilentLocalPairing({
-        locality,
-        hasBrowserOriginHeader: false,
-        isControlUi: false,
-        isWebchat: false,
-        isNativeAppUi: true,
-        reason: "metadata-upgrade",
-      }),
-    ).toBe(true);
-    expect(
-      shouldAllowSilentLocalPairing({
-        locality: "remote",
-        hasBrowserOriginHeader: false,
-        isControlUi: false,
-        isWebchat: false,
-        isNativeAppUi: true,
-        reason: "not-paired",
-      }),
-    ).toBe(false);
-  });
-
   it("isolates browser-origin loopback clients in the browser limiter", () => {
     const rateLimiter = createRateLimiter();
     const browserRateLimiter = createRateLimiter();
@@ -225,7 +172,7 @@ describe("handshake auth helpers", () => {
 
   it("recommends device-token retry only for shared-token mismatch with device identity", () => {
     const resolved = resolveUnauthorizedHandshakeContext({
-      connectAuth: { token: "test-token" },
+      connectAuth: { token: "shared-token" },
       failedAuth: { ok: false, reason: "token_mismatch" },
       hasDeviceIdentity: true,
     });
@@ -239,7 +186,7 @@ describe("handshake auth helpers", () => {
 
   it("treats explicit device-token mismatch as credential update guidance", () => {
     const resolved = resolveUnauthorizedHandshakeContext({
-      connectAuth: { deviceToken: "test-token-device" },
+      connectAuth: { deviceToken: "device-token" },
       failedAuth: { ok: false, reason: "device_token_mismatch" },
       hasDeviceIdentity: true,
     });
@@ -253,7 +200,7 @@ describe("handshake auth helpers", () => {
 
   it("treats device-token scope mismatch as configuration review guidance", () => {
     const resolved = resolveUnauthorizedHandshakeContext({
-      connectAuth: { deviceToken: "test-token-device" },
+      connectAuth: { deviceToken: "device-token" },
       failedAuth: { ok: false, reason: "scope_mismatch" },
       hasDeviceIdentity: true,
     });
